requirements.txt
ollama>=0.5.0
pymupdf>=1.26.0
pillow>=10.0.0
tiktoken>=0.8.0 
langchain>=0.1.0
langchain-community>=0.0.10
pypdf2>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
python-docx>=0.8.11
docx2txt>=0.8 